virtual_saree_tryon/requirements.txt
streamlit==1.37.0
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.1
mediapipe==0.10.14
